# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.Response-any:
    properties:
      data: {}
      error:
        type: string
      total:
        type: integer
    type: object
...
  handler.Response-handler_uploadResponse:
    properties:
      data:
        $ref: '#/definitions/handler.uploadResponse'
      error:
        type: string
      total:
        type: integer
    type: object
  handler.uploadFileResult:
    properties:
      error:
        type: string
      file:
        type: string
      id:
        type: string
    type: object
  handler.uploadResponse:
    properties:
      message:
        type: string
      results:
        items:
          $ref: '#/definitions/handler.uploadFileResult'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for receiving, processing and classifying NF-e (Nota Fiscal Eletrônica)
    XML files.
  title: NF-e Processor API
  version: "1.0"
paths:
...
  /xml/upload:
    post:
      consumes:
      - multipart/form-data
      description: Receives one or more NF-e XML files, validates and enqueues for
        async processing via RabbitMQ
      parameters:
      - description: XML file(s)
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.Response-handler_uploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response-any'
      summary: Upload NF-e XML files
      tags:
      - nfe
schemes:
- http
- https
swagger: "2.0"